main: ProxiedHost.buildUrl() can just slice the src path to set the dst path
Since we ensure with ProxiedHost.addHandle() that the handler will
always match on paths like /NAME/, where NAME is ProxiedHost.name.
Thus by just using Path[len("/NAME"):], we effectively get the rest
of the string with a "/" trailing (so it is a valid path).

And since len("/NAME") has O(1) complexity, it is faster than the
previous implementation.